Fix bug in computing gradient penalty
This is a shameful bug... Obviously the part I edited was copied directly from the Github repo of improved training for Wasserstein GANs. However, in the author's code, the image was first flatten to a matrix, that's why the reduction axis was only for [1]. In this implementation, we should use reduction axis [1, 2, 3] because we don't flatten images.